Use custom format string.
In a recent change I introduced a regression because
I thought using `--format=oneline --abbrev-commit` gave
me the exact same output as much custom format string.

This was not the case. I introduced a regression in my
string parsing of the results, because --oneline literally
means one line per commit. But, my parsing was allowing for
blank lines between commits. So this change goes back
to my custom format string.